requirements.txt
Flask==3.0.0
opencv-python==4.9.0.80
numpy==1.26.2
Pillow==10.1.0
google-cloud-storage==2.13.0
google-cloud-aiplatform==1.38.1
python-dotenv==1.0.0
werkzeug==3.0.1
scikit-learn==1.3.2
tensorflow==2.15.0
gdown==4.7.1
matplotlib==3.8.2
google-api-python-client==2.111.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas>=2.2.0
kagglehub>=0.3.0
plotly>=5.18.0
dash>=2.15.0
seaborn>=0.13.0
joblib>=1.3.0 